pyproject.toml
[tool.poetry]
name = "ccaconfig"
version = "0.1.0"
description = ""
authors = ["ccdale <chris.allison@hivehome.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.2"
pytest = "^5.3.2"
ccalogging = "^0.3.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
